# src/main/java/com/recette/apk/Models/Recette.java
package com.recette.apk.models;

import jakarta.persistence.*;
import lombok.Data;
import lombok.NoArgsConstructor;

@Entity
@Data
@NoArgsConstructor
public class Recette {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    private String nom;
    private String ingredients;
    private Double prix;
    private String pays;
    private String saison;
}